rename gc to gc/v1
