+++ pkg/vm/engine/tae/db/gc/v1/diskcleaner.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gc
+package v1
 
 import (
 	"context"
